Rename pidee to pidee-cli (see commit message)
Debatable, but yes, there's `redis-cli` which talks to Redis server, but then again Node's `npm-cli.js` is linked as `/bin/npm`, so I don't know if there's a "correct" way of naming this based on the patterns I know.
Most command line utlities don't end in `-cli`, but things that have a Command Line Interface counterpart, have a `-cli` executable. Now should we consider `pidee-service` as the "main" Pidee thing, even though it's never called directly by the user? Or should we give the privilege of a clean name to what is now called `pidee-cli` and just go for `pidee`?
As you can understand from the above, I'm fifty/fifty about this. But I do think that as trivial as it sounds it's importnat as it will contribute to the identity of this project somewhat. Would love to hear your thoughts.